refactor(docs): コンポーネントページを動的ルートからmdxに移行
