+++ app/ginga-ui.com/src/app/components/accordion/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  MultipleExpansionExample,
+  DefaultClosedExample,
+  ControlledExample,
+} from "#/examples/accordion";
+
+<ComponentDoc id="accordion">
+  <Example id="accordion" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="accordion" name="MultipleExpansionExample">
+    <MultipleExpansionExample />
+  </Example>
+  <Example id="accordion" name="DefaultClosedExample">
+    <DefaultClosedExample />
+  </Example>
+  <Example id="accordion" name="ControlledExample">
+    <ControlledExample />
+  </Example>
+</ComponentDoc>
