+++ app/ginga-ui.com/src/app/components/anchor/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  ExternalExample,
+  DisabledExample,
+  ButtonVariantExample,
+} from "#/examples/anchor";
+
+<ComponentDoc id="anchor">
+  <Example id="anchor" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="anchor" name="ExternalExample">
+    <ExternalExample />
+  </Example>
+  <Example id="anchor" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="anchor" name="ButtonVariantExample">
+    <ButtonVariantExample />
+  </Example>
+</ComponentDoc>
