+++ app/ginga-ui.com/src/app/components/box/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  LayoutExample,
+  CardStyleExample,
+  NestedExample,
+} from "#/examples/box";
+
+<ComponentDoc id="box">
+  <Example id="box" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="box" name="LayoutExample">
+    <LayoutExample />
+  </Example>
+  <Example id="box" name="CardStyleExample">
+    <CardStyleExample />
+  </Example>
+  <Example id="box" name="NestedExample">
+    <NestedExample />
+  </Example>
+</ComponentDoc>
