+++ app/ginga-ui.com/src/app/components/button/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  VariantsExample,
+  DisabledExample,
+  FormExample,
+} from "#/examples/button";
+
+<ComponentDoc id="button">
+  <Example id="button" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="button" name="VariantsExample">
+    <VariantsExample />
+  </Example>
+  <Example id="button" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="button" name="FormExample">
+    <FormExample />
+  </Example>
+</ComponentDoc>
