+++ app/ginga-ui.com/src/app/components/card/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  WithFooterExample,
+  MultipleCardsExample,
+  SimpleExample,
+} from "#/examples/card";
+
+<ComponentDoc id="card">
+  <Example id="card" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="card" name="WithFooterExample">
+    <WithFooterExample />
+  </Example>
+  <Example id="card" name="MultipleCardsExample">
+    <MultipleCardsExample />
+  </Example>
+  <Example id="card" name="SimpleExample">
+    <SimpleExample />
+  </Example>
+</ComponentDoc>
