+++ app/ginga-ui.com/src/app/components/checkbox/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  SelectedExample,
+  DisabledExample,
+  CheckboxGroupExample,
+} from "#/examples/checkbox";
+
+<ComponentDoc id="checkbox">
+  <Example id="checkbox" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="checkbox" name="SelectedExample">
+    <SelectedExample />
+  </Example>
+  <Example id="checkbox" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="checkbox" name="CheckboxGroupExample">
+    <CheckboxGroupExample />
+  </Example>
+</ComponentDoc>
