+++ app/ginga-ui.com/src/app/components/dialog/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  ConfirmationExample,
+  FormExample,
+  AlertExample,
+} from "#/examples/dialog";
+
+<ComponentDoc id="dialog">
+  <Example id="dialog" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="dialog" name="ConfirmationExample">
+    <ConfirmationExample />
+  </Example>
+  <Example id="dialog" name="FormExample">
+    <FormExample />
+  </Example>
+  <Example id="dialog" name="AlertExample">
+    <AlertExample />
+  </Example>
+</ComponentDoc>
