+++ app/ginga-ui.com/src/app/components/form-control/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  WithIdExample,
+  MultipleFieldsExample,
+  FormExample,
+} from "#/examples/form-control";
+
+<ComponentDoc id="form-control">
+  <Example id="form-control" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="form-control" name="WithIdExample">
+    <WithIdExample />
+  </Example>
+  <Example id="form-control" name="MultipleFieldsExample">
+    <MultipleFieldsExample />
+  </Example>
+  <Example id="form-control" name="FormExample">
+    <FormExample />
+  </Example>
+</ComponentDoc>
