+++ app/ginga-ui.com/src/app/components/heading/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  AllLevelsExample,
+  PageStructureExample,
+  CustomStyleExample,
+} from "#/examples/heading";
+
+<ComponentDoc id="heading">
+  <Example id="heading" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="heading" name="AllLevelsExample">
+    <AllLevelsExample />
+  </Example>
+  <Example id="heading" name="PageStructureExample">
+    <PageStructureExample />
+  </Example>
+  <Example id="heading" name="CustomStyleExample">
+    <CustomStyleExample />
+  </Example>
+</ComponentDoc>
