+++ app/ginga-ui.com/src/app/components/image/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  AvatarExample,
+  SizesExample,
+  RoundedExample,
+} from "#/examples/image";
+
+<ComponentDoc id="image">
+  <Example id="image" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="image" name="AvatarExample">
+    <AvatarExample />
+  </Example>
+  <Example id="image" name="SizesExample">
+    <SizesExample />
+  </Example>
+  <Example id="image" name="RoundedExample">
+    <RoundedExample />
+  </Example>
+</ComponentDoc>
