+++ app/ginga-ui.com/src/app/components/input/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  TypesExample,
+  DisabledExample,
+  WithValueExample,
+} from "#/examples/input";
+
+<ComponentDoc id="input">
+  <Example id="input" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="input" name="TypesExample">
+    <TypesExample />
+  </Example>
+  <Example id="input" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="input" name="WithValueExample">
+    <WithValueExample />
+  </Example>
+</ComponentDoc>
