+++ app/ginga-ui.com/src/app/components/list/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  UnorderedExample,
+  OrderedExample,
+  NestedExample,
+  MixedExample,
+} from "#/examples/list";
+
+<ComponentDoc id="list">
+  <Example id="list" name="UnorderedExample">
+    <UnorderedExample />
+  </Example>
+  <Example id="list" name="OrderedExample">
+    <OrderedExample />
+  </Example>
+  <Example id="list" name="NestedExample">
+    <NestedExample />
+  </Example>
+  <Example id="list" name="MixedExample">
+    <MixedExample />
+  </Example>
+</ComponentDoc>
