+++ app/ginga-ui.com/src/app/components/paragraph/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  MultipleExample,
+  LongTextExample,
+  CustomStyleExample,
+} from "#/examples/paragraph";
+
+<ComponentDoc id="paragraph">
+  <Example id="paragraph" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="paragraph" name="MultipleExample">
+    <MultipleExample />
+  </Example>
+  <Example id="paragraph" name="LongTextExample">
+    <LongTextExample />
+  </Example>
+  <Example id="paragraph" name="CustomStyleExample">
+    <CustomStyleExample />
+  </Example>
+</ComponentDoc>
