+++ app/ginga-ui.com/src/app/components/radio/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  DefaultValueExample,
+  DisabledExample,
+  ControlledExample,
+} from "#/examples/radio";
+
+<ComponentDoc id="radio">
+  <Example id="radio" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="radio" name="DefaultValueExample">
+    <DefaultValueExample />
+  </Example>
+  <Example id="radio" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="radio" name="ControlledExample">
+    <ControlledExample />
+  </Example>
+</ComponentDoc>
