+++ app/ginga-ui.com/src/app/components/select/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  DisabledOptionsExample,
+  DisabledExample,
+  ControlledExample,
+} from "#/examples/select";
+
+<ComponentDoc id="select">
+  <Example id="select" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="select" name="DisabledOptionsExample">
+    <DisabledOptionsExample />
+  </Example>
+  <Example id="select" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="select" name="ControlledExample">
+    <ControlledExample />
+  </Example>
+</ComponentDoc>
