+++ app/ginga-ui.com/src/app/components/slider/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  CustomRangeExample,
+  WithStepsExample,
+  DisabledExample,
+} from "#/examples/slider";
+
+<ComponentDoc id="slider">
+  <Example id="slider" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="slider" name="CustomRangeExample">
+    <CustomRangeExample />
+  </Example>
+  <Example id="slider" name="WithStepsExample">
+    <WithStepsExample />
+  </Example>
+  <Example id="slider" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+</ComponentDoc>
