+++ app/ginga-ui.com/src/app/components/switch/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  SelectedExample,
+  DisabledExample,
+  ControlledExample,
+} from "#/examples/switch";
+
+<ComponentDoc id="switch">
+  <Example id="switch" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="switch" name="SelectedExample">
+    <SelectedExample />
+  </Example>
+  <Example id="switch" name="DisabledExample">
+    <DisabledExample />
+  </Example>
+  <Example id="switch" name="ControlledExample">
+    <ControlledExample />
+  </Example>
+</ComponentDoc>
