+++ app/ginga-ui.com/src/app/components/tab/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  VerticalExample,
+  DisabledTabExample,
+  ControlledExample,
+} from "#/examples/tab";
+
+<ComponentDoc id="tab">
+  <Example id="tab" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="tab" name="VerticalExample">
+    <VerticalExample />
+  </Example>
+  <Example id="tab" name="DisabledTabExample">
+    <DisabledTabExample />
+  </Example>
+  <Example id="tab" name="ControlledExample">
+    <ControlledExample />
+  </Example>
+</ComponentDoc>
