+++ app/ginga-ui.com/src/app/components/table/page.mdx
@@ -1,0 +1,22 @@
+import { ComponentDoc, Example } from "#/components/component-doc";
+import {
+  BasicExample,
+  WithSelectionExample,
+  UserTableExample,
+  EmptyStateExample,
+} from "#/examples/table";
+
+<ComponentDoc id="table">
+  <Example id="table" name="BasicExample">
+    <BasicExample />
+  </Example>
+  <Example id="table" name="WithSelectionExample">
+    <WithSelectionExample />
+  </Example>
+  <Example id="table" name="UserTableExample">
+    <UserTableExample />
+  </Example>
+  <Example id="table" name="EmptyStateExample">
+    <EmptyStateExample />
+  </Example>
+</ComponentDoc>
